validação da remoção de categoria

+++ src/pages/admin/components/categories/actions.js
@@ -9,6 +9,15 @@ export async function getCategories() {
   })
 }
 
+export async function getWords() {
+  const db = firebase.firestore();
+  const words = await db.collection('words').get();
+
+  return words.docs.map(doc => {
+    return { id: doc.id, ...doc.data() }
+  })
+}
+
 export async function addCategory(name, callback) {
   const db = firebase.firestore();
   await db.collection('categories').add({
@@ -19,6 +28,9 @@ export async function addCategory(name, callback) {
 };
 
 export async function removeCategory(id, callback) {
+  if (await hasCategoryInWord(id)) {
+    return;
+  }
   const db = firebase.firestore();
   await db.collection('categories').doc(id).delete();
   callback();
@@ -32,4 +44,13 @@ export async function updateCategory(id, name, callback) {
     name: name
   });
   callback();
+}
+
+export async function hasCategoryInWord(idCategory) {
+  let hasCategoryInWord = false
+  const wordArray = await getWords();
+  const categoryArray = await getCategories();
+  const categorySelected = categoryArray.filter(category => { if (category.id == idCategory) return category });
+  wordArray.map(word => { if (word.category.name == categorySelected[0].name) hasCategoryInWord = true })
+  return hasCategoryInWord;
 }
